style: add explicit typing and use cast for type safety, update mypy ignores
- Add typing and cast usage across modules for improved type safety
- Update function signatures and variable annotations to use explicit types
- Refactor list/dict initializations and return types for clarity
- Add or adjust mypy: ignore-errors comments in test files
- Minor logic tweaks to ensure type correctness and suppress mypy warnings

+++ tests/unit/test_metrics.py
@@ -1,3 +1,4 @@
+# mypy: ignore-errors
 import pandas as pd
 import pytest
 
@@ -38,7 +39,9 @@ def test_max_drawdown_known_value() -> None:
 
 def test_calmar_positive() -> None:
     # Needs a drawdown to produce a non-zero calmar
-    values = [10_000 + i * 5 for i in range(126)] + [10_000 + i * 10 for i in range(126)]
+    values = [float(10_000 + i * 5) for i in range(126)] + [
+        float(10_000 + i * 10) for i in range(126)
+    ]
     eq = _equity(values)
     assert calmar(eq) > 0
